chore: Change frameType from byte to short.

+++ codec-http2/src/main/java/io/netty5/handler/codec/http2/Http2EventAdapter.java
@@ -82,7 +82,7 @@ public class Http2EventAdapter implements Http2Connection.Listener, Http2FrameLi
     }
 
     @Override
-    public void onUnknownFrame(ChannelHandlerContext ctx, byte frameType, int streamId, Http2Flags flags,
+    public void onUnknownFrame(ChannelHandlerContext ctx, short frameType, int streamId, Http2Flags flags,
                                Buffer payload) throws Http2Exception {
     }
