+++ codec-http2/src/test/java/io/netty5/handler/codec/http2/Http2FrameCodecTest.java
@@ -243,10 +243,10 @@ public class Http2FrameCodecTest {
         Http2ConnectionDecoder dec = new DefaultHttp2ConnectionDecoder(conn, enc, new DefaultHttp2FrameReader());
         new Http2FrameCodec(enc, dec, new Http2Settings(), false, true) {
             @Override
-            protected Http2StreamFrame newHttp2UnknownFrame(byte frameType,
-                                                      int streamId,
-                                                      Http2Flags flags,
-                                                      Buffer payload) {
+            protected Http2StreamFrame newHttp2UnknownFrame(short frameType,
+                                                            int streamId,
+                                                            Http2Flags flags,
+                                                            Buffer payload) {
                 return super.newHttp2UnknownFrame(frameType, streamId, flags, payload);
             }
         };
